+++ packages/openai/src/responses/openai-responses-api-types.ts
@@ -1,17 +1,16 @@
 import { JSONSchema7 } from '@ai-sdk/provider';
 
-export type OpenAIResponsesPrompt = Array<OpenAIResponsesMessage>;
+export type OpenAIResponsesInput = Array<OpenAIResponsesInputItem>;
 
-export type OpenAIResponsesMessage =
+export type OpenAIResponsesInputItem =
   | OpenAIResponsesSystemMessage
   | OpenAIResponsesUserMessage
   | OpenAIResponsesAssistantMessage
   | OpenAIResponsesFunctionCall
   | OpenAIResponsesFunctionCallOutput
-  | OpenAIWebSearchCall
-  | OpenAIComputerCall
-  | OpenAIFileSearchCall
-  | OpenAIResponsesReasoning;
+  | OpenAIResponsesComputerCall
+  | OpenAIResponsesReasoning
+  | OpenAIResponsesItemReference;
 
 export type OpenAIResponsesIncludeOptions =
   | Array<
@@ -45,12 +44,7 @@ export type OpenAIResponsesUserMessage = {
 
 export type OpenAIResponsesAssistantMessage = {
   role: 'assistant';
-  content: Array<
-    | { type: 'output_text'; text: string }
-    | OpenAIWebSearchCall
-    | OpenAIComputerCall
-    | OpenAIFileSearchCall
-  >;
+  content: Array<{ type: 'output_text'; text: string }>;
   id?: string;
 };
 
@@ -68,22 +62,15 @@ export type OpenAIResponsesFunctionCallOutput = {
   output: string;
 };
 
-export type OpenAIWebSearchCall = {
-  type: 'web_search_call';
-  id: string;
-  status?: string;
-};
-
-export type OpenAIComputerCall = {
+export type OpenAIResponsesComputerCall = {
   type: 'computer_call';
   id: string;
   status?: string;
 };
 
-export type OpenAIFileSearchCall = {
-  type: 'file_search_call';
+export type OpenAIResponsesItemReference = {
+  type: 'item_reference';
   id: string;
-  status?: string;
 };
 
 export type OpenAIResponsesTool =